# backend_fashion_app/internal/common/httpx/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  resp.Envelope:
    properties:
      code:
        type: string
      data: {}
      message:
        type: string
    type: object
info:
  contact: {}
  description: Auth (local, Google) + Profile (customer/shop) + Catalog
  title: MyFashion API
  version: "1.0"
paths:
...
  /profiles/me:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/resp.Envelope'
      security:
      - Bearer: []
      summary: Get my profile
      tags:
      - Profile
    put:
      consumes:
      - multipart/form-data
      description: Tạo hoặc cập nhật hồ sơ. Hỗ trợ upload avatar qua file (multipart/form-data)
        hoặc URL (form field).
      parameters:
      - description: (Customer) Tên đầy đủ
        in: formData
        name: full_name
        type: string
      - description: (Customer) Tuổi
        in: formData
        name: age
        type: integer
      - description: (Customer) Giới tính
        in: formData
        name: gender
        type: string
      - description: Địa chỉ (chung cho Customer và Shop)
        in: formData
        name: address
        type: string
      - description: (Shop) Tên cửa hàng
        in: formData
        name: shop_name
        type: string
      - description: File ảnh avatar để tải lên
        in: formData
        name: avatar
        type: file
      - description: URL ảnh avatar (nếu không tải file)
        in: formData
        name: img_url
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Envelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/resp.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/resp.Envelope'
      security:
      - Bearer: []
      summary: Upsert my profile
      tags:
      - Profile
...
schemes:
- http
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"